requirements.txt
pandas~=1.1.2
python-dotenv~=0.14.0
psycopg2
SQLAlchemy~=1.3.19

requests~=2.24.0
Unidecode~=1.1.1
wikipedia~=1.4.0
beautifulsoup4~=4.9.1
selenium~=3.141.0